Rename algorithm/alg -> stl/algorithm.h.
Also change the namesapce from alg:: to stl::.

In preparation for introducing additional STL polyfills.

GPC: move-algorithm
