+++ controller/lib/hal/hal.h
@@ -35,7 +35,7 @@ limitations under the License.
 // observe whether mocked methods are called.  So far that hasn't been
 // necessary.
 
-#include "alg.h"
+#include "algorithm.h"
 
 #ifdef TEST_MODE
 
@@ -356,14 +356,14 @@ inline void HalApi::analogWrite(PwmPin pin, int value) {
   ::analogWrite(rawPin(pin), value);
 }
 [[nodiscard]] inline uint16_t HalApi::serialRead(char *buf, uint16_t len) {
-  return Serial.readBytes(buf, alg::min(len, serialBytesAvailableForRead()));
+  return Serial.readBytes(buf, stl::min(len, serialBytesAvailableForRead()));
 }
 inline uint16_t HalApi::serialBytesAvailableForRead() {
   return Serial.available();
 }
 [[nodiscard]] inline uint16_t HalApi::serialWrite(const char *buf,
                                                   uint16_t len) {
-  return Serial.write(buf, alg::min(len, serialBytesAvailableForWrite()));
+  return Serial.write(buf, stl::min(len, serialBytesAvailableForWrite()));
 }
 inline uint16_t HalApi::serialBytesAvailableForWrite() {
   return Serial.availableForWrite();
@@ -430,7 +430,7 @@ inline void HalApi::analogWrite(PwmPin pin, int value) {
     return 0;
   }
   auto &readBuf = serialIncomingData_.front();
-  uint16_t n = alg::min<uint16_t>(len, readBuf.size());
+  uint16_t n = stl::min<uint16_t>(len, readBuf.size());
   memcpy(buf, readBuf.data(), n);
   readBuf.erase(readBuf.begin(), readBuf.begin() + n);
   if (readBuf.empty()) {
@@ -443,7 +443,7 @@ inline uint16_t HalApi::serialBytesAvailableForRead() {
 }
 [[nodiscard]] inline uint16_t HalApi::serialWrite(const char *buf,
                                                   uint16_t len) {
-  uint16_t n = alg::min(len, serialBytesAvailableForWrite());
+  uint16_t n = stl::min(len, serialBytesAvailableForWrite());
   serialOutgoingData_.insert(serialOutgoingData_.end(), buf, buf + n);
   return n;
 }
@@ -453,7 +453,7 @@ inline uint16_t HalApi::serialBytesAvailableForWrite() {
   return 64;
 }
 inline uint16_t HalApi::test_serialGetOutgoingData(char *data, uint16_t len) {
-  uint16_t n = alg::min<uint16_t>(len, serialOutgoingData_.size());
+  uint16_t n = stl::min<uint16_t>(len, serialOutgoingData_.size());
   memcpy(data, serialOutgoingData_.data(), n);
   serialOutgoingData_.erase(serialOutgoingData_.begin(),
                             serialOutgoingData_.begin() + n);
