+++ controller/lib/stl/algorithm.cpp
